requirements.txt
runpod==1.7.0
huggingface_hub==0.26.5
pillow==10.4.0
